fix(blinds) Allow blinds to reconnect if network is absent at startup.
The situation is the following when we turn on electricity in the
house:

- The router/switch is starting,
- The blind microcontroller is starting and tries to connect.

But the router takes more time to start than the blind
microcontroller. So the blind controller enters in an error state.

This patch fixes that by allowing the blind microcontroller to retry
to connect to the network 5 times, with a delay of 15 seconds between
each try, before entering the error state.