fix(lights) Allow lights to reconnect if network is absent at startup.
The situation is the following when we turn on electricity in the
house:

- The router/switch is starting,
- The light microcontroller is starting and tries to connect.

But the router takes more time to start than the light
microcontroller. So the light controller enters in an error state.

This patch fixes that by allowing the light microcontroller to retry
to connect to the network 5 times, with a delay of 15 seconds between
each try, before entering the error state.

It's similar to 1c2d32d624c1cc880ca093ff645b856e10ee96f2.