feat(ui) Reduce network request by fetching all property values at once.
Instead of reading each property individually, the `read_property`
function is now replaced by `fetch_properties` which can receive a
list of property to fetch. Once this function returns, the user has to
call the `read_properties` function, which will refresh the properties
values at once, and then will propose all the properties details.

For a cold startup, it reduces the number of network requests from 21
to 14. And for a basic usage, from 15 to 8.